pyproject.toml
[project]
name = "modelx-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "black>=23.0.0",
    "bs4>=0.0.2",
    "chromadb>=1.3.5",
    "dagshub>=0.6.3",
    "fastapi>=0.122.0",
    "fasttext-wheel>=0.9.2",
    "flake8>=6.0.0",
    "groq>=0.11.0",
    "hdbscan>=0.8.40",
    "joblib>=1.5.2",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-neo4j>=0.6.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.4.7",
    "lingua-language-detector>=2.1.1",
    "lxml>=5.0.0",
    "mlflow>=3.7.0",
    "neo4j>=5.28.2",
    "numpy>=2.3.5",
    "optuna>=4.6.0",
    "pandas>=2.0.0",
    "playwright>=1.56.0",
    "plotly>=5.18.0",
    "pydantic>=2.0.0",
    "pymongo>=4.15.5",
    "pypdf>=6.4.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "pytz>=2024.1",
    "pyyaml>=6.0.3",
    "requests>=2.31.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.28.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "umap-learn>=0.5.9.post2",
    "uvicorn[standard]>=0.38.0",
    "watchdog>=6.0.0",
    "websockets>=15.0.1",
    "yfinance>=0.2.0",
]
